# implement vectorized evaluation for builtinCastTimeAsTimeSig