@ekalinin 2019-11-08
I'm sorry, but `builtinCastTimeAsTimeSig` has been vectorized by #13209