@XuHuaiyu 2019-11-11
@chengmyy Hi, thanks for your contribution. It will be nice if you can choose another function to make it vectorized.

@ekalinin Thanks for the informing.